修复错误

+++ model/utils/functional.py
@@ -2,7 +2,7 @@ import torch
 
 
 def inverse_sigmoid(x, eps: float = 1e-5):
-    x = x.clamp(min=0, max=1)
+    x = x.to(torch.float).clamp(min=0, max=1)
     x1 = x.clamp(min=eps)
     x2 = (1 - x).clamp(min=eps)
     return torch.log(x1 / x2)
